+++ types/rpc.go
@@ -27,8 +27,9 @@ type AccountInfo struct {
 
 //Estimate represents estimated gas will be used and storage will be collateralized when transaction excutes
 type Estimate struct {
-	GasLimit *hexutil.Big `json:"gasLimit"`
-	GasUsed  *hexutil.Big `json:"gasUsed"`
+	GasLimit              *hexutil.Big `json:"gasLimit"`
+	GasUsed               *hexutil.Big `json:"gasUsed"`
+	StorageCollateralized *hexutil.Big `json:"storageCollateralized"`
 }
 
 type RewardInfo struct {